% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaPlasticScore_pkg.R
\name{MetaPlasticScore}
\alias{MetaPlasticScore}
\title{MetaPlasticScorePipeline.R}
\description{
Pipeline for computing a plastic degradation potential Score
score from HMMER (hmmsearch) outputs and taxon abundance tables.
}
\details{
The file contains parsers for HMMER output, matrix builders, scoring logic,
statistical tests and plotting (ggplot2 + ComplexHeatmap).
}
\author{
Gaetano Aufiero

Maintainer: Gaetano Aufiero
}
